feat: Update display options for hex encoding in test-captain4
This commit updates the display options for hex encoding in the `test-captain4` project. It introduces a new constant `DISPLAY_HEX` in the `definition.rs` file, which controls whether the hex encoding should be displayed or not. By default, the hex encoding is now disabled. This change improves the readability of the output and reduces unnecessary information when running the program.