feat: Increase transaction amount and include data in sendTransaction
This commit increases the transaction amount from 0.001 ETH to 0.1 ETH and includes 32 bytes of string data in the sendTransaction method. The data is encoded using ethers.encodeBytes32String and logged using console.log. This change improves the functionality of the transaction and allows for the inclusion of additional information in the transaction.