feat: Add AlexNet model implementation and test script
This commit adds the implementation of the AlexNet model in the `test-alexnet/a.py` file. The model consists of several convolutional and fully connected layers. It also includes a `forward` method for performing forward pass and a `predict` method for making predictions with softmax probabilities. Additionally, a test script `test_alexnet` is added to demonstrate the usage of the model.

New file: test-alexnet/a.py
New function: test_alexnet